split lhadaptor into its own file
